+++ src/sample_demo/06TestConstract.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.25;
 
 contract Counter {
-    uint public count;
+    uint256 public count;
     address public add;
 
     function getCount() public {
