+++ src/sample_demo/09-2TestArray.sol
@@ -3,19 +3,19 @@ pragma solidity 0.8.25;
 
 contract TestArray {
     // 容易理解错误：表示分配一个初始空间，不是代表第一个元素为 1
-    uint[] public arr = new uint[](1);
+    uint256[] public arr = new uint256[](1);
 
     function addWithoutValue() public {
         // 用来添加新的零初始化元素到数组末尾，并返回元素的引用，以便修改元素的内容，如：x.push().t = 2或x.push() = b，push方法只对存储（storage）中的动态数组有效。
         arr.push();
     }
 
-    function addWithValue(uint x) public {
+    function addWithValue(uint256 x) public {
         // 用来添加给定元素到数组末尾。push(x) 没有返回值，方法只对存储（storage）中的动态数组有效
         arr.push(x);
     }
 
-    function getLen() public view returns (uint) {
+    function getLen() public view returns (uint256) {
         return arr.length;
     }
 
@@ -24,20 +24,19 @@ contract TestArray {
         arr.pop();
     }
 
-    function getArr() public view returns (uint[] memory) {
+    function getArr() public view returns (uint256[] memory) {
         return arr;
     }
-
 }
 
 contract ArrayGas {
-    uint[] numbers;
-    uint sum;
-    uint currentIndex;
+    uint256[] numbers;
+    uint256 sum;
+    uint256 currentIndex;
 
     // 分段计算
-    function calculateSum(uint endIndex) public {
-        if(endIndex > currentIndex) {
+    function calculateSum(uint256 endIndex) public {
+        if (endIndex > currentIndex) {
             for (uint256 i = currentIndex; i < endIndex; i++) {
                 sum += numbers[i];
             }
@@ -45,13 +44,13 @@ contract ArrayGas {
     }
 
     // 推荐数组移除元素的实现
-    function remove(uint index) public {
-        uint len = numbers.length;
+    function remove(uint256 index) public {
+        uint256 len = numbers.length;
         if (index == len - 1) {
             numbers.pop();
         } else {
             // 被删除元素与最后一个元素交换位置
-            numbers[index] = numbers[len -1];
+            numbers[index] = numbers[len - 1];
             numbers.pop();
         }
     }
@@ -60,11 +59,11 @@ contract ArrayGas {
 contract TestStringBytes {
     bytes bs;
     bytes bs0 = "12abcd";
-    bytes bs1 = "abc\x22\x22";   // 十六进制数
-    bytes bs2 = "Alan\u661f";   //u661f为汉字“星”的Unicode编码值
+    bytes bs1 = "abc\x22\x22"; // 十六进制数
+    bytes bs2 = "Alan\u661f"; //u661f为汉字“星”的Unicode编码值
 
     string str = "AlanYin";
-    
+
     string name;
 
     function setName(string calldata _name) public {
