+++ src/sample_demo/17TestTryCatch.sol
@@ -8,40 +8,39 @@ pragma solidity 0.8.25;
 // • 注意：out of gas 错误不是程序异常，错误不能捕获。
 
 interface IFoo {
-    function myFunc1(uint x) external pure returns (uint);
-    function myFunc2(uint x) external pure returns (uint);
+    function myFunc1(uint256 x) external pure returns (uint256);
+    function myFunc2(uint256 x) external pure returns (uint256);
 }
 
 contract Foo {
-
     error NotZero();
 
-    function myFunc1(uint x) public pure returns (uint) {
+    function myFunc1(uint256 x) public pure returns (uint256) {
         // require(x > 0, "x must gt 0");
-        if ( x == 0) revert NotZero();
+        if (x == 0) revert NotZero();
 
         return x + 1;
     }
 }
 
 contract TestTryCatch {
-
     IFoo public foo;
-    uint public x;
+    uint256 public x;
 
     string public errString;
     bytes public errBytes;
 
-    constructor (address _foo) {
+    constructor(address _foo) {
         foo = IFoo(_foo);
     }
 
-    function TryCatchExternalCall(uint _i) public {
+    function TryCatchExternalCall(uint256 _i) public {
         foo.myFunc1(_i);
-        
-        try foo.myFunc2(_i) returns (uint result) {
+
+        try foo.myFunc2(_i) returns (uint256 result) {
             x = result;
-        } catch Error(string memory reason) { //  catch revert 
+        } catch Error(string memory reason) {
+            //  catch revert
             errString = reason;
         } catch (bytes memory reason) {
             errBytes = reason;
