+++ src/sample_demo/20TestInterface.sol
@@ -3,11 +3,11 @@ pragma solidity 0.8.25;
 
 interface ICounter {
     function increment() external;
-    function count() external view returns (uint);
+    function count() external view returns (uint256);
 }
 
 contract Counter {
-    uint public count;
+    uint256 public count;
 
     function increment() public {
         count += 1;
@@ -19,7 +19,7 @@ contract TestUse {
         ICounter(_counter).increment();
     }
 
-    function getCount(address _counter) public view returns (uint) {
+    function getCount(address _counter) public view returns (uint256) {
         return ICounter(_counter).count();
     }
 }
